Fixed JPanel repainting issues updated Game View
Made it so the menu page actually displays instead of only showing a
blank JFrame. It was one line of code that took me a god damn hour to
figure out and fix. Also updated GameView class to display an undo
button. Tried to get it to work with a layout manager but failed. Unless
someone else can figure it out we may have to manually enter in the
coordinates for the game. Which would be fine since we would have to
anyway to add the ellipses that contain the playing stones.